Fix issues when updating insights
