+++ src/SIL.XForge.Scripture/ClientApp/src/app/translate/editor/lynx/insights/lynx-insight-editor-objects/lynx-insight-editor-objects.component.ts
@@ -1,13 +1,15 @@
 import { Component, DestroyRef, Input, OnDestroy, OnInit } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { isEqual } from 'lodash-es';
-import { filter, merge, switchMap, tap } from 'rxjs';
-import { pairwise } from 'rxjs/operators';
+import { combineLatest, filter, fromEvent, merge, switchMap, tap } from 'rxjs';
+import { map, pairwise } from 'rxjs/operators';
 import { EditorReadyService } from '../base-services/editor-ready.service';
 import { InsightRenderService } from '../base-services/insight-render.service';
 import { LynxableEditor } from '../lynx-editor';
+import { LynxInsight, LynxInsightDisplayState, LynxInsightRange } from '../lynx-insight';
 import { LynxInsightOverlayService } from '../lynx-insight-overlay.service';
 import { LynxInsightStateService } from '../lynx-insight-state.service';
+import { LynxInsightBlot } from '../quill-services/blots/lynx-insight-blot';
 
 @Component({
   selector: 'app-lynx-insight-editor-objects',
@@ -15,6 +17,10 @@ import { LynxInsightStateService } from '../lynx-insight-state.service';
   styleUrl: './lynx-insight-editor-objects.component.scss'
 })
 export class LynxInsightEditorObjectsComponent implements OnInit, OnDestroy {
+  readonly insightSelector = `.${LynxInsightBlot.superClassName}`;
+
+  private readonly dataIdProp = LynxInsightBlot.idDatasetPropName;
+
   @Input() editor?: LynxableEditor;
 
   constructor(
@@ -29,6 +35,17 @@ export class LynxInsightEditorObjectsComponent implements OnInit, OnDestroy {
     if (this.editor == null) {
       return;
     }
+
+    combineLatest([
+      fromEvent(this.editor, 'selection-change').pipe(map(([range]) => range)),
+      this.insightState.filteredChapterInsights$
+    ])
+      .pipe(takeUntilDestroyed(this.destroyRef))
+      .subscribe(([range, insights]) => this.handleSelectionChange(range, insights));
+
+    combineLatest([fromEvent(this.editor.root, 'mouseover'), this.insightState.filteredChapterInsights$])
+      .pipe(takeUntilDestroyed(this.destroyRef))
+      .subscribe(([event]) => this.handleMouseOver(event.target as HTMLElement));
 
     this.editorReadyService
       .listenEditorReadyState(this.editor)
@@ -86,4 +103,60 @@ export class LynxInsightEditorObjectsComponent implements OnInit, OnDestroy {
       this.insightRenderService.removeAllInsightFormatting(this.editor);
     }
   }
+
+  private handleSelectionChange(selection: LynxInsightRange | undefined, insights: LynxInsight[]): void {
+    console.log('SelectionChange', selection, insights);
+    const ids = insights
+      .filter(insight => selection != null && overlaps(insight.range, selection))
+      .map(insight => insight.id);
+
+    let displayStateChanges: Partial<LynxInsightDisplayState> = {
+      activeInsightIds: ids,
+      promptActive: ids.length > 0,
+      actionOverlayActive: false
+    };
+
+    this.insightState.updateDisplayState(displayStateChanges);
+  }
+
+  private handleMouseOver(target: HTMLElement): void {
+    // Clear any 'hover-insight' classes if the target is not an insight element
+    if (!target.matches('.' + LynxInsightBlot.superClassName)) {
+      this.insightState.updateDisplayState({ cursorActiveInsightIds: [] });
+      return;
+    }
+
+    console.log('MouseOver', target);
+    const ids: string[] = this.getInsightIds(target);
+
+    // Set 'hover-insight' class on the affected insight elements (clear others)
+    this.insightState.updateDisplayState({ cursorActiveInsightIds: ids });
+  }
+
+  /**
+   * Get all insight ids from the element and its parents that match the lynx insight selector.
+   */
+  private getInsightIds(el: HTMLElement): string[] {
+    const ids: string[] = [];
+
+    if (el.matches(this.insightSelector)) {
+      let currentEl: HTMLElement | null | undefined = el;
+
+      while (currentEl != null) {
+        const id = currentEl.dataset[this.dataIdProp];
+
+        if (id != null) {
+          ids.push(id);
+        }
+
+        currentEl = currentEl.parentElement?.closest(this.insightSelector);
+      }
+    }
+
+    return ids;
+  }
+}
+
+function overlaps(x: LynxInsightRange, y: LynxInsightRange): boolean {
+  return x.index <= y.index + y.length && y.index <= x.index + x.length;
 }
